Add verify list is deleted
I added a few steps to verify that the play list is deleted.  Checks if an existing playlist called "Ricardo" exits and if it does not, it creates it and then deletes and checks for deletion.  This code does not work if an existing playlist called "Ricardo" already exists.  I do not undertand why